styles: Fix line-height of delete button of image-upload-widget.
This commit fixes the bug with the delete button of both profile
picture and realm-logo buttons. The button was shifted downwards
and was caused due to line height change in 24a5bca442a4.

Found the correct value of 20px by checking the line-height in
one test organization on zulipchat.com where the button position
looks correct.