embed_links: Interrupt consume() function on worker timeout.
This fixes a bug introduced in 0453029ab8c28b11431ff5ef684ac6e680ab2bb0
which made the worker simply log a warning about the timeout and then
continue consume()ing the event that should have also been interrupted.

The idea here is to introduce an exception which can be used to
interrupt the consume() process without triggering the regular handling
of exceptions that happens in _handle_consume_exception.