android notifications: Differentiate personal vs wildcard mentions.
The code to also notify for wildcard mentions was added in
6b86ad321e15fb7285f59bc623cc1a150cf3f77f.

But that showed the same text for both the cases. This commit fixes
that.

This is more of change for correctness. The mobile app currently does
not rely on this text for notifications, but constructs the text by
itself from the data in the payload.

This also fixes the "stream_push_notify" case to consistently show
a `#` before the stream name.