composebox_typeahead: Trigger foucs to send button only after keyup event.
As explained in the commit 6ff7c2a330ed2204dff55ec5ab794f86dd470b33:
`nextFocus` is set on a keydown event to indicate where
we should focus on keyup. We can't focus at the time of
keydown because we need to wait for typeahead. And we
can't compute where to focus at the time of keyup because
only the keydown has reliable information about whether it
was a Tab or a Shift+Tab.

Though the previous configuration didn't show any bug but we should
follow the same logic here also.