settings: Fix incorrect elements used for emojiset loading spinner.
We were using emojiset_spinner element in destroy_indicator, but
there is no such element and emoji-settings-status element is used
in make_indicator, so destroy_indicator should also use the same
element.

This was added originally in ebe9f87cf7c5 (though in a different
file) and was removed for sometime and then readded in 3be1339592.
But between these two commits, the setting was changed to be a
radio element and thus the spinner elements were also changed.