settings: Remove unnecessary code.
We were using incorrect element as we changed the input from
select to radio buttons in 8d7ea49fae9.
This code was originally added in ebe9f87cf7c5 (though in a
different file) and was removed from sometime and then added
again in 3be1339592b. But in the meantime the input was
changed to radio, which resulted in incorrect code.