email notifs: Update scheduled timestamps after batching period change.
The API for changing the batching period was added in
fd1b9e6d8fb40a801167321785b934a59ccae6d3.

This is a follow up to that commit. We also update the timestamps for
existing scheduled email notifications entries so that the effect of
changing the setting is immediate.

Part of #15280