user_groups: Hide system user groups from UI.
We do not have any system user groups as of
now, but this commit is just a prep commit
to prevent any change in user-facing pages
to avoid confusion till this feature is
completed.

This change was initially made in b73b1df,
but it was reverted in 42f2bf43 due to merge
conflicts with the typescript migration of
user_groups.js.