openapi: Use more carefully-constructed regexes for markdown extension.
This removes a false-positive ReDoS, since the input is always
checked-in code.  It also incidentally refactors to make the regexes
be more explicit about the values they expect, and removes unnecessary
capturing groups.

It removes an optional parenthesized status code for fixtures,
unnecessary since 25991630130953af8118451fbce5cd63948df7f8, as well as
optional key-value language options, unnecessary since
8a1052ed8bf41c0b73cbb1fa6c4a887172c87f1f.

Thank you to @erik-krogh and @yoff for bringing this to our attention.