stream settings: Fix exception for users who cannot create streams.
The previous logic, introduced in
3b4a2def35e85c32e6eeff7f1ae4b354548c828b, crashed when visiting
"manage streams" in the case that the user didn't have permission to
create any streams at all.

We remove the more complex logic.

This isn't the right long term fix, because it means users who can
only create public streams just get a vague "Insufficient permission"
error when trying to create a private stream (and vice versa), but is
sufficient to fix the critical bug of "manage streams" failing to load.

Fixes #19895.