message_view_header: Fix hover effect related to links in description.
This handler was broken during refactor
72faab1dc84e51afd782d3ded9e0b07025105377, as we can see from the
original implementation in f7e4befa29e1c66a868ad58c5c6453b313b54091,
the intent is that hovering over any link within the
narrow_description should not cause the search_icon to change color ie
the hover effect should not be used. This is so because it aligns
with the fact that clicking the links would not open the search bar.

However, during the refactor this was incorrectly switched to forcing
the effect to be applied when we hover over links in the
narrow_description.

This commit reverts to the original and intended behaviour, and also
switches to using opacity rather than color, in accordance with the
changes from the previous commit
(316d499ac74c2caddb57c98a43d9b776b1b32d98).