queue_processors: Drop unused current_queue_size, which was local size.
The `current_queue_size` key in the queue monitoring stats file was
the local queue size, not the global queue size -- a955a91892a2
renamed the function, but did not adjust the queue monitoring JSON,
despite the last use of it having been removed in ddb0e20b5602.

The function is still used to mark "we emptied our queue," and it
remains a reasonable metric for that.